requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.30.0
safetensors>=0.3.0

# Data & Evaluation
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API Clients
openai>=1.0.0
anthropic>=0.20.0
groq>=0.4.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
